+++ Editor/Path/PathEditor.cs
@@ -1,4 +1,6 @@
 using Etienne;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEditor;
 using UnityEditorInternal;
 using UnityEngine;
@@ -7,122 +9,161 @@ namespace EtienneEditor {
     [CustomEditor(typeof(Path))]
     public class PathEditor : Editor<Path> {
 
-        private void OnEnable() {
-            SetupPathList();
+        private SerializedProperty waypointsProperty;
+        private List<int> selectedIndex = new List<int>();
+
+        private void OnEnable()
+        {
+            waypointsProperty = serializedObject.FindProperty("waypoints");
         }
 
-        private void SetupPathList() {
-            waypointsList = new ReorderableList(serializedObject, serializedObject.FindProperty("waypoints")) {
-                drawElementCallback = (Rect rect, int index, bool isActive, bool isFocused) => DrawElement(rect, index),
-                drawHeaderCallback = (Rect rect) => EditorGUI.LabelField(rect, "Path"),
-                onCanRemoveCallback = (ReorderableList list) => waypointsList.count > 2,
-                onAddCallback = (ReorderableList list) => AddElement(list)
-            };
-        }
-
-        private void AddElement(ReorderableList list) {
-            int index = list.serializedProperty.arraySize;
-            list.serializedProperty.arraySize++;
-            list.index = index;
-            SerializedProperty element = list.serializedProperty.GetArrayElementAtIndex(index);
-            Vector3[] lasts = new Vector3[2] {
-                list.serializedProperty.GetArrayElementAtIndex(waypointsList.index - 1).vector3Value,
-                list.serializedProperty.GetArrayElementAtIndex(waypointsList.index - 2).vector3Value };
-            Vector3 pos = (lasts[0] - lasts[1]) + lasts[0];
-            element.vector3Value = pos;
-        }
-
-        private void DrawElement(Rect rect, int index) {
-            SerializedProperty element = waypointsList.serializedProperty.GetArrayElementAtIndex(index);
-            rect.y += 2;
-            EditorGUI.PropertyField(new Rect(rect.x, rect.y, rect.width, EditorGUIUtility.singleLineHeight),
-                element, GUIContent.none);
-        }
-
-        public override void OnInspectorGUI() {
-            if(waypointsList == null) {
-                SetupPathList();
+        private void OnSceneGUI()
+        {
+            HandleOutOfRange();
+            HandleShortcuts();
+            using (new Handles.DrawingScope(Color.white, Target.transform.localToWorldMatrix))
+            {
+                DrawCurve();
+                DrawPositionHandles();
+                DrawButtons();
             }
-
-            serializedObject.Update();
-            EditorGUI.BeginChangeCheck();
-            EditorGUILayout.PropertyField(serializedObject.FindProperty("isLooped"));
-            EditorGUILayout.LabelField("Path Lenght", Target.PathLenght.ToString("0.00"));
-            waypointsList.DoLayoutList();
-            if(EditorGUI.EndChangeCheck()) {
-                serializedObject.ApplyModifiedProperties();
-            }
+            serializedObject.ApplyModifiedProperties();
         }
 
-        private void OnSceneGUI() {
-            Color oldColor = Handles.color;
-            if(Tools.current == Tool.Move) {
-                Matrix4x4 localToWorld = Target.transform.localToWorldMatrix;
-                for(int i = 0; i < waypointsList.count; i++) {
-                    DrawSelectionHandle(i, localToWorld);
-                    if(waypointsList.index == i) {
-                        DrawPositionControl(i, localToWorld, Target.transform.rotation);
-                    }
+        private void HandleShortcuts()
+        {
+            Event e = Event.current;
+            if (e == null) return;
+            if (e.isKey && e.type == EventType.KeyDown && selectedIndex.Count > 0)
+            {
+                if (e.control && e.keyCode == KeyCode.D)
+                {
+                    DuplicatePoints();
+                    e.Use();
+                }
+                if (e.keyCode == KeyCode.Delete)
+                {
+                    DeletePoints();
+                    e.Use();
                 }
             }
-            DrawPath();
-            Handles.color = oldColor;
         }
 
-        private void DrawPath() {
-            Handles.color = Color.white;
-            Handles.DrawAAPolyLine(Target.WorldWaypoints);
+        private void DeletePoints()
+        {
+            selectedIndex.Sort();
+            selectedIndex = selectedIndex.Distinct().ToList();
+            if (waypointsProperty.arraySize <= 2) return;
+            for (int i = selectedIndex.Count - 1; i >= 0; i--)
+            {
+                waypointsProperty.DeleteArrayElementAtIndex(selectedIndex[i]);
+            }
+            for (int i = selectedIndex.Count - 1; i >= 0; i--)
+            {
+                int newValue = selectedIndex[i] - (i + 1);
+                if (newValue < 0) continue;
+                else selectedIndex[i] = newValue;
+            }
+            selectedIndex.Sort();
+            selectedIndex = selectedIndex.Distinct().ToList();
         }
 
-        private void DrawPositionControl(int i, Matrix4x4 localToWorld, Quaternion localRotation) {
-            Handles.color = Color.white;
-            Vector3 waypoint = Target.LocalWaypoints[i];
-            EditorGUI.BeginChangeCheck();
-            Vector3 pos = localToWorld.MultiplyPoint(waypoint);
-            Quaternion rotation = (Tools.pivotRotation == PivotRotation.Local) ? localRotation : Quaternion.identity;
-            float size = HandleUtility.GetHandleSize(pos) * .1f;
-            Handles.SphereHandleCap(0, pos, rotation, size, EventType.Repaint);
-            pos = Handles.PositionHandle(pos, rotation);
-            if(EditorGUI.EndChangeCheck()) {
-                Undo.RecordObject(target, "Move Waypoint");
-                waypoint = Matrix4x4.Inverse(localToWorld).MultiplyPoint(pos);
-                serializedObject.FindProperty("waypoints").GetArrayElementAtIndex(i).vector3Value = waypoint;
-                serializedObject.ApplyModifiedProperties();
-                ForceSceneUpdate();
+        private void DuplicatePoints()
+        {
+            selectedIndex.Sort();
+            selectedIndex = selectedIndex.Distinct().ToList();
+            for (int i = selectedIndex.Count - 1; i >= 0; i--)
+            {
+                waypointsProperty.InsertArrayElementAtIndex(selectedIndex[i]);
+            }
+            for (int i = selectedIndex.Count - 1; i >= 0; i--)
+            {
+                selectedIndex[i] += i + 1;
+            }
+            selectedIndex.Sort();
+            selectedIndex = selectedIndex.Distinct().ToList();
+        }
+
+        private void HandleOutOfRange()
+        {
+            for (int i = selectedIndex.Count - 1; i >= 0; i--)
+            {
+                if (selectedIndex[i] >= waypointsProperty.arraySize)
+                {
+                    selectedIndex.Remove(selectedIndex[i]);
+                    Debug.Log($"Removing index {i}");
+                }
             }
         }
 
-        private void DrawSelectionHandle(int i, Matrix4x4 localToWorld) {
-            Handles.color = Color.white;
-            Vector3 pos = localToWorld.MultiplyPoint(Target.LocalWaypoints[i]);
-
-            float size = HandleUtility.GetHandleSize(pos) * 0.2f;
-
-            if(Handles.Button(pos, Quaternion.identity, size, size, Handles.SphereHandleCap) && waypointsList.index != i) {
-                waypointsList.index = i;
-                ForceSceneUpdate();
+        private void DrawPositionHandles()
+        {
+            SerializedProperty position;
+            Vector3 delta = Vector3.zero;
+            Quaternion rotation = Quaternion.identity;
+            for (int i = 0; i < selectedIndex.Count; i++)
+            {
+                position = waypointsProperty.GetArrayElementAtIndex(selectedIndex[i]);
+                delta += Handles.PositionHandle(position.vector3Value, rotation) - position.vector3Value;
             }
-
-            Handles.BeginGUI();
-
-            Vector2 labelSize = Vector2.one * EditorGUIUtility.singleLineHeight;
-            Vector2 labelPos = HandleUtility.WorldToGUIPoint(pos);
-            labelPos.x -= labelSize.x / 2;
-            labelPos.y -= labelSize.y / 2;
-
-            GUILayout.BeginArea(new Rect(labelPos, labelSize));
-
-            GUIStyle style = new GUIStyle() {
-                alignment = TextAnchor.MiddleCenter
-            };
-            style.normal.textColor = Color.black;
-            GUILayout.Label(new GUIContent(i.ToString(), "Waypoint " + i), style);
-
-            GUILayout.EndArea();
-
-            Handles.EndGUI();
+            for (int i = 0; i < selectedIndex.Count; i++)
+            {
+                position = waypointsProperty.GetArrayElementAtIndex(selectedIndex[i]);
+                position.vector3Value += delta;
+            }
         }
 
-        private ReorderableList waypointsList;
+        private void DrawButtons()
+        {
+            SerializedProperty position;
+            Quaternion rotation = Quaternion.identity;
+            for (int i = 0; i < waypointsProperty.arraySize; i++)
+            {
+                position = waypointsProperty.GetArrayElementAtIndex(i);
+                float size = HandleUtility.GetHandleSize(position.vector3Value) * 0.2f;
+                Color color = selectedIndex.Contains(i) ? (Color)EtienneEditor.EditorColor.Orange : Color.white;
+
+                using (new Handles.DrawingScope(color))
+                {
+                    if (Handles.Button(position.vector3Value, rotation, size, size*.5f, Handles.SphereHandleCap))
+                    {
+                        if (!Event.current.control) selectedIndex.Clear();
+                        if (selectedIndex.Contains(i)) selectedIndex.Remove(i);
+                        else selectedIndex.Add(i);
+                    }
+                }
+
+                Handles.BeginGUI();
+
+                Vector2 labelSize = Vector2.one * EditorGUIUtility.singleLineHeight;
+                Vector2 labelPos = HandleUtility.WorldToGUIPoint(position.vector3Value);
+                labelPos.x -= labelSize.x / 2;
+                labelPos.y -= labelSize.y / 2;
+
+                GUILayout.BeginArea(new Rect(labelPos, labelSize));
+
+                GUIStyle style = new GUIStyle()
+                {
+                    alignment = TextAnchor.MiddleCenter
+                };
+                style.normal.textColor = Color.black;
+                GUILayout.Label(new GUIContent(i.ToString(), "Waypoint " + i), style);
+
+                GUILayout.EndArea();
+
+                Handles.EndGUI();
+            }
+        }
+
+        private void DrawCurve()
+        {
+            Handles.DrawAAPolyLine(Target.CatmullWaypoints);
+            Color color = Handles.color;
+            color.a = .2f;
+            using (new Handles.DrawingScope(color, Handles.matrix))
+            {
+                Handles.DrawAAPolyLine(Target.LocalWaypoints);
+            }
+        }
     }
 }